backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pillow==10.3.0
captcha==0.5.0
python-magic==0.4.27
slowapi==0.1.9
redis==5.0.4
aioboto3==13.1.1
aiofiles==23.2.1
httpx==0.27.0
python-dotenv==1.0.1
email-validator==2.1.1
